Bug 5094 - Code cleanup, patch by Zack Middleton and DevHC. Fixes unused-but-set gcc warnings Added also specific Smokin'Guns code cleanup
(cherry picked from commit 23f6fd1633465c2f750f7b1c490e18b0cffa2cd8)

+++ code/server/sv_ccmds.c
@@ -1762,7 +1762,7 @@ void SVCL_SaveRecord_f( void ) {
 		return;
 	}
 	if ( !cl->demorecording ) {
-		Com_Printf ("Not recording a demo for client %i.\n", cl - svs.clients);
+		Com_Printf ("Not recording a demo for client %i.\n", (int)(cl - svs.clients));
 		return;
 	}
 	clientnum = cl - svs.clients;
